feat: Enhance Text2SQL API with SQL generation endpoint and request/response models
