+++ src/utils.py
@@ -137,3 +137,102 @@ def df_to_json(df: pd.DataFrame) -> str:
     """Convert pandas DataFrame to JSON (records format)."""
     return df.to_json(orient="records", indent=2)
 
+def build_prompt(user_question: str) -> str:
+    return f"""
+        You are a helpful assistant specializing in data analysis in a PostgreSQL warehouse.
+        Answer the questions by providing SQL code that is compatible with the PostgreSQL environment.
+        This is the question you are required to answer:
+        {user_question}
+
+        Here is the relevant context of the database:
+        -- Countries
+        CREATE TABLE countries (
+            country_id SERIAL PRIMARY KEY,
+            country_name VARCHAR(100) NOT NULL UNIQUE
+        );
+
+        -- Regions
+        CREATE TABLE regions (
+            region_id SERIAL PRIMARY KEY,
+            region_name VARCHAR(100) NOT NULL,
+            country_id INT NOT NULL REFERENCES countries(country_id)
+        );
+
+        -- Cities
+        CREATE TABLE cities (
+            city_id SERIAL PRIMARY KEY,
+            city_name VARCHAR(100) NOT NULL,
+            region_id INT NOT NULL REFERENCES regions(region_id)
+        );
+
+        -- Customers
+        CREATE TABLE customers (
+            customer_id VARCHAR(10) PRIMARY KEY,
+            company_name VARCHAR(100) NOT NULL,
+            contact_name VARCHAR(100),
+            contact_title VARCHAR(50),
+            address TEXT,
+            city_id INT REFERENCES cities(city_id),
+            postal_code VARCHAR(20),
+            phone VARCHAR(30),
+            fax VARCHAR(30)
+        );
+
+        -- Employees
+        CREATE TABLE employees (
+            employee_id SERIAL PRIMARY KEY,
+            last_name VARCHAR(50) NOT NULL,
+            first_name VARCHAR(50) NOT NULL,
+            title VARCHAR(50),
+            title_of_courtesy VARCHAR(25),
+            birth_date DATE,
+            hire_date DATE,
+            city_id INT REFERENCES cities(city_id)
+        );
+
+        -- Suppliers
+        CREATE TABLE suppliers (
+            supplier_id SERIAL PRIMARY KEY,
+            company_name VARCHAR(100) NOT NULL,
+            city_id INT REFERENCES cities(city_id)
+        );
+
+        -- Categories
+        CREATE TABLE categories (
+            category_id SERIAL PRIMARY KEY,
+            category_name VARCHAR(100) NOT NULL UNIQUE
+        );
+
+        -- Products
+        CREATE TABLE products (
+            product_id SERIAL PRIMARY KEY,
+            product_name VARCHAR(100) NOT NULL,
+            supplier_id INT REFERENCES suppliers(supplier_id),
+            category_id INT REFERENCES categories(category_id)
+        );
+
+        -- Shippers
+        CREATE TABLE shippers (
+            shipper_id SERIAL PRIMARY KEY,
+            company_name VARCHAR(100) NOT NULL
+        );
+
+        -- Orders
+        CREATE TABLE orders (
+            order_id SERIAL PRIMARY KEY,
+            customer_id VARCHAR(10) REFERENCES customers(customer_id),
+            employee_id INT REFERENCES employees(employee_id),
+            order_date DATE,
+            ship_via INT REFERENCES shippers(shipper_id)
+        );
+
+        -- Order Details
+        CREATE TABLE order_details (
+            order_id INT NOT NULL REFERENCES orders(order_id),
+            product_id INT NOT NULL REFERENCES products(product_id),
+            unit_price NUMERIC(10,2) NOT NULL,
+            quantity SMALLINT NOT NULL,
+            discount REAL DEFAULT 0,
+            PRIMARY KEY (order_id, product_id)
+        );
+        """
